requirements.txt
fastapi uvicorn
langchain-groq
langchain-community
langchain-core
faiss-cpu
python-dotenv
sentence-transformers